Adding a passthrough encoder and new app server. Also removed all old files in favor of the new structure.
