+++ package.json
@@ -13,7 +13,7 @@
     "dotenv": "^7.0.0",
     "esm": "^3.2.16",
     "mqtt": "^2.14.0",
-    "websocket": "^1.0.25"
+    "ws": "^7.4.2"
   },
   "engines": {
     "node": ">= 11.0.0",
