--- src/animations/index.js
@@ -1,11 +1,0 @@
-// Legacy
-
-import { RainbowScroll } from './rainbow_scroll';
-import { ColorLadder } from './color_ladder';
-import { Snowfall } from './snowfall';
-
-export default {
-  'rainbow_scroll': RainbowScroll,
-  'color_ladder': ColorLadder,
-  'snowfall': Snowfall,
-};
