--- src/animations/snowfall.js
@@ -1,27 +1,0 @@
-import { AnimationBase } from './base';
-import { rgb } from '../color';
-
-export class Snowfall extends AnimationBase {
-  init() {
-    console.log('RainbowScroll -> start');
-
-    // let snow_sprites = [];
-
-    this.config = {
-      brightness: 100,
-      speed: 50,
-      intensity: 50
-    };
-
-    // Override some internal config variables
-    this.interval = 1000 / 60;
-
-    // Storage Variables
-    this.step = 0;
-  }
-
-  frame() {
-    this.fill(rgb(200, 200, 200));
-    this.step += this.config.speed;
-  }
-}
