--- src/services/server_conn.js
@@ -1,15 +1,0 @@
-// Device Discovery Service
-//
-// This service uses [SSDP | MDNS] to discover IP connected
-// Aurora Devices across your local network. Once discovered
-// the devices will be addedd to the device store.
-
-export class ServerConnService {
-  constructor() {
-    
-  }
-
-  listen() {
-    
-  }
-}
